Refactor transfer copy helpers
